requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
billiard==4.2.1
celery==5.5.0rc4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.6
django-celery-beat==2.7.0
django-timezone-field==7.1
djangorestframework==3.15.2
drf-spectacular==0.28.0
et_xmlfile==2.0.0
flake8==7.1.2
idna==3.10
inflection==0.5.1
isort==6.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.0rc3
lxml==5.3.1
marshmallow==3.26.1
marshmallow_dataclass==8.7.1
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.2.3
openpyxl==3.2.0b1
packaging==24.2
pandas==2.2.3
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.3.0b5
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typeguard==4.4.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
